% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as_robonomist_datasources}
\alias{as_robonomist_datasources}
\title{Create a robonomist_datasources object for tidy printing}
\usage{
as_robonomist_datasources(x)
}
\arguments{
\item{x}{A data.frame with columns \code{id}, \code{title}, and \code{langugages}}
}
\description{
Create a robonomist_datasources object for tidy printing
}
\keyword{internal}
