% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robonomist_id.R
\name{as_robonomist_id}
\alias{as_robonomist_id}
\title{Cast object into a robonomist id}
\usage{
as_robonomist_id(x)
}
\arguments{
\item{x}{object to be cast}
}
\description{
Cast object into a robonomist id
}
