% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robonomist_id.R
\name{$.robonomist_id}
\alias{$.robonomist_id}
\title{Extract components of robonomist_id with $-operator}
\usage{
\method{$}{robonomist_id}(x, ...)
}
\arguments{
\item{x}{robonomist_id}

\item{...}{for compatibility}
}
\description{
Extract components of robonomist_id with $-operator
}
