% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{data_get}
\alias{data_get}
\title{Get data for table id}
\usage{
data_get(id, dl_filter = NULL, labels = TRUE, lang = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{id}{string. Exact table id}
}
\description{
Returns the objects without performing searching or pattern filters. Fails if no match exists.
}
