% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{data_metadata}
\alias{data_metadata}
\title{Metadata of table}
\usage{
data_metadata(id)
}
\arguments{
\item{String, }{exact id of table}
}
\value{
List of metadata
}
\description{
Currently works only with px-based tables
}
