% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{data_vintage}
\alias{data_vintage}
\title{Vintage of a data table}
\usage{
data_vintage(id)
}
\arguments{
\item{id}{Character vector, Exact table ids}
}
\value{
Named vector of POSIXct-class datetimes indicating the vintage of each table.
}
\description{
\code{data_vintage()} returns the latest update time of the specified data tables for the purpose of monitoring data freshness.

The vintage indicates the time when the table of was last updated at the source, if such information is provided by the datasource. If the source does not provide vintage information, the vintage will be the time new data was last queried from the source.
}
