% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{datasource_menu}
\alias{datasource_menu}
\title{Structured menu of available datasources and data tables}
\usage{
datasource_menu(datasource = NULL, lang = NULL)
}
\arguments{
\item{datasource, }{Character, Datasource name}

\item{lang, }{Two-letter language code, e.g. "en" or "sv".}
}
\value{
Nested list of datasource_menu and datasource_menu_item objects
}
\description{
Structured menu of available datasources and data tables
}
