% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{datasources}
\alias{datasources}
\title{List available datasources}
\usage{
datasources(lang = NULL, available_only = TRUE)
}
\arguments{
\item{lang, }{Two-letter language code, e.g. "en" or "sv".}

\item{available_only, }{If TRUE (default), then list only those datasources which are available for the current subscription.}
}
\description{
List available datasources
}
