% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_data_from_url}
\alias{fetch_data_from_url}
\title{Parse and Retrieve Data from URL}
\usage{
fetch_data_from_url(url, get = TRUE)
}
\arguments{
\item{url}{A character string representing the URL copied from the datasource's website when a particular dataset is selected.}

\item{get}{A logical value indicating whether to execute the query and return the data (\code{TRUE}, default) or just return the query string (\code{FALSE}).}
}
\value{
If \code{get} is \code{TRUE}, the function returns the queried data. If \code{get} is \code{FALSE}, it returns the query string.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because \code{\link[=data]{data()}} function now supports direct data fetching from various datasources.

The \code{fetch_data_from_url()} function is designed to facilitate easy access to data by parsing a copied URL from websites of various datasources. It forms a query to the database and retrieves the data in R using the \code{robonomistclient} package.

Currently supported sites:
\itemize{
\item \href{https://data-explorer.oecd.org/}{OECD Data Explorer}
\item \href{https://ec.europa.eu/eurostat/databrowser/explore/all/all_themes}{Eurostat Data Explorer}
\item \href{https://data.ecb.europa.eu/}{ECB Data Portal}
\item \href{https://data.worldbank.org/}{World Bank}
\item \href{https://fred.stlouisfed.org/}{Fred Economic Data}
\item \href{https://data.bis.org/}{Bank of International Settlements, BIS}
\item \href{https://stat.fi/tup/tilastotietokannat/index.html}{Databases of Statistics Finland}
\item \href{https://tilastot.etk.fi/pxweb/en/ETK}{Database of Finnish Centre for Pensions}
\item \href{https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/}{Statistics Sweden}
}
}
\examples{
# Example usage:
url <- "https://data-explorer.oecd.org/vis?tm=sna&pg=0&fs[0]=Measure\%2C0\%7CAquaculture\%20production\%23AQUA_PD\%23&fc=Measure&snb=1&vw=tb&df[ds]=dsDisseminateFinalDMZ&df[id]=DSD_FISH_PROD\%40DF_FISH_AQUA&df[ag]=OECD.TAD.ARP&df[vs]=1.0&pd=2010\%2C&dq=.A.._T.T&ly[rw]=REF_AREA&ly[cl]=TIME_PERIOD&to[TIME_PERIOD]=false"
fetch_data_from_url(url)
fetch_data_from_url(url, get = FALSE)
# StatFin
fetch_data_from_url("https://statfin.stat.fi/PxWeb/pxweb/fi/StatFin/StatFin__aku/statfin_aku_pxt_12ea.px/")

# Eurostat fetch_data_from_url("https://ec.europa.eu/eurostat/databrowser/view/cens_hnmga/default/table?lang=en&category=cens.cens_hn.cens_hnstr")

# ECB
fetch_data_from_url("https://data.ecb.europa.eu/data/datasets/ICP/ICP.M.U2.N.000000.4.ANR")

# World bank
fetch_data_from_url("https://data.worldbank.org/indicator/SH.DYN.MORT?locations=1W&start=1990&view=chart")

# Fred
fetch_data_from_url("https://fred.stlouisfed.org/series/FPCPITOTLZGUSA")

# BIS
fetch_data_from_url("https://data.bis.org/topics/RPP/BIS,WS_SPP,1.0/Q.5R.N.628")

# SCB
fetch_data_from_url("https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__EN__EN0302/SSDArGasavtal/")

}
