% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_oecd_data_from_url}
\alias{fetch_oecd_data_from_url}
\title{Parse and Retrieve OECD Data from URL}
\usage{
fetch_oecd_data_from_url(url, get = TRUE)
}
\arguments{
\item{url}{A character string representing the URL copied from the OECD Data Explorer when a particular dataset is selected.}

\item{get}{A logical value indicating whether to execute the query and return the data (\code{TRUE}, default) or just return the query string (\code{FALSE}).}
}
\value{
If \code{get} is \code{TRUE}, the function returns the queried data. If \code{get} is \code{FALSE}, it returns the query string.
}
\description{
The \code{fetch_oecd_data_from_url()} function is designed to facilitate easy access to OECD data by parsing a copied URL from the \href{https://data-explorer.oecd.org/}{OECD Data Explorer}. It forms a query to the database and retrieves the data in R using the \code{robonomistclient} package.
}
\examples{
# Example usage:
url <- "https://data-explorer.oecd.org/vis?tm=sna&pg=0&fs[0]=Measure\%2C0\%7CAquaculture\%20production\%23AQUA_PD\%23&fc=Measure&snb=1&vw=tb&df[ds]=dsDisseminateFinalDMZ&df[id]=DSD_FISH_PROD\%40DF_FISH_AQUA&df[ag]=OECD.TAD.ARP&df[vs]=1.0&pd=2010\%2C&dq=.A.._T.T&ly[rw]=REF_AREA&ly[cl]=TIME_PERIOD&to[TIME_PERIOD]=false"
data <- fetch_oecd_data_from_url(url)
query_string <- fetch_oecd_data_from_url(url, get = FALSE)
}
