% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogle.R
\name{ogle}
\alias{ogle}
\title{Quickly explore categorical and summary variables in a tibble}
\usage{
ogle(.data, ...)
}
\arguments{
\item{.data}{A tibble to explore.}

\item{...}{<'data-masking'> Optional. Variables to include in the exploration. If omitted, all variables are explored.}
}
\value{
A list: for character and factor variables, their unique values; for numeric and date/time variables, a summary.
}
\description{
Provides a concise overview of a tibble by listing unique values for categorical variables and summary statistics for numeric or date/time variables. Useful for initial data exploration before filtering or analysis.
}
\examples{
\dontrun{
data("sotkanet/4") \%>\% ogle()
data("sotkanet/4") \%>\% ogle(region)
}
}
