% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{data}
\alias{data}
\alias{data_get}
\alias{data_search}
\title{Search and get data from Robonomist Data Server}
\usage{
dаta(
  pattern = "",
  dl_filter = NULL,
  labels = getOption("robonomistClient.labels"),
  lang = NULL,
  na.rm = FALSE,
  tidy_time = getOption("robonomistClient.tidy_time"),
  ...
 )

data_get(
  id,
  dl_filter = NULL,
  labels = getOption("robonomistClient.labels"),
  lang = NULL,
  na.rm = FALSE,
  tidy_time = getOption("robonomistClient.tidy_time"),
  ...
)

data_search(pattern = "")
}
\arguments{
\item{pattern}{Character. Search query or table id, possibly followed by a \code{§}-filter.}

\item{dl_filter}{List or named vector. The download filter is passed to datasource download functions to filter data before download. See Details.}

\item{labels}{Logical. If FALSE, then variable values are returned as codes instead of labels.}

\item{lang}{Two-letter language code, e.g. "en" or "sv".}

\item{na.rm}{Logical. If TRUE, and supported by the datasource, missing values are removed from the returned data frame upon retrieval.}

\item{tidy_time}{Logical. If TRUE, the time dimension is parsed into Date class and renamed \code{time}. If NULL, the datasource specific default will be used.}

\item{...}{Other arguments passed to datasource download functions.}

\item{id, }{Character. Exact table id. If multiple ids are provided as a character vector, they are bound row-wise into a single data frame.}
}
\description{
\code{data()} is a convenience function that searches the database and returns
\enumerate{
\item search results if multiple matching tables are found, and
\item data if only one match is found.
}

\code{data_get()} returns data for a given table id without performing searching, url parsing or pattern filters, and is prefered in programmatic use-cases where robustness and speed are important. \code{data_get()} will result in error, if no match exists.

\code{data_search()} performs a search and returns a list of matching data tables without downloading any data.
}
\section{\verb{§}-filter}{
The \code{data()} function allows for a special \code{§}-filter. When the pattern matches a single table and the function returns a data frame, the \code{§}-filter can be used to subset rows of data frame. The tibble is filterd by sequence of regular expressions separated by section sign \code{§}. The regular expression are applied to data frame's variables sequentially.

If the last variable is a date, it is used as a start date filter.
}

\section{Download filter}{
Some datasources (e.g. datasets "ecb" & "tulli") do not allow downloading full data tables, nor is it always preferred due to the large large size of table. For these datasources the user must provide a download filter via the \code{dl_filter} argument. When the argument is left as \code{NULL}, the  \code{data} function will return a list of variables and potential values. This list can be used to construct a suitable download filter.

Generally, \code{dl_filter} should be named list where names are variable names and values character vectors of selected values (see Examples). Alternatively, some datasources allow for a dot-separated string to define a download filter.
}

\examples{

# Search for datasets matching pattern:
data("consumer indicator")

## Limit your search to a specific dataset by providing
## the dataset name and a slash as prefix:
data("ec/ consumer indicator")

## Download data by providing exact table id:
data("ec/consumer")

## With time series based datasets you can retrieve the time series
## using the source's id:
data("ecb/FM.M.U2.EUR.RT.MM.EURIBOR1YD_.HSTA")

## Alternatively, you can copy the full URL from the source's website:
data("https://data.ecb.europa.eu/data/datasets/FM/FM.M.U2.EUR.RT.MM.EURIBOR1YD_.HSTA")

## Most time series datasets also support wildcards. For example,
## in case of ECB, you can leave a part of the series id unspecified:
data("ecb/FM.M.U2.EUR.RT.MM..HSTA")

## If the data table too large to download in full, you may need to
## provide a download filter. First get the available variables and values:
data("ecb/AME") |> str()

## Then provide a suitable filter to download data:
data("ecb/AME", dl_filter = list(ame_ref_area = "FIN"))

## Another example with Finish Customs dataset:
data("tulli/uljas_cpa2008",
  dl_filter = list(
    "Tavaraluokitus CPA2008_2" = "*A-X",
    "Aika" = c("201505", "201506"),
    "Maa" = "=ALL",
    "Suunta" = "=FIRST 1",
    "Indikaattorit" = "=FIRST 1"
  )
)

## Using §-filter to filter data after download:
data("ec/consumer§Fin§Confidence")

## Using §-filter with start date:
data("ec/consumer§Fin§Confidence§2020-01-01")

}
