% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{data_metadata}
\alias{data_metadata}
\title{Metadata of table}
\usage{
data_metadata(id, lang = NULL)
}
\arguments{
\item{id, }{Exact table id}

\item{lang, }{Two-letter language code, e.g. "en" or "sv".}
}
\value{
List of metadata
}
\description{
\code{data_metadata()} returns metadata information about a specific data table. The returned metadata structure depends on the datasource.
}
