% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{data_metadata}
\alias{data_metadata}
\title{Metadata of table}
\usage{
data_metadata(id, lang = NULL)
}
\arguments{
\item{id, }{Exact table id}
}
\value{
List of metadata
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The output of this function is likely to change. Currently works only with px-based tables.
}
