% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{data_vintage}
\alias{data_vintage}
\title{Vintage of table, i.e. datetime of the latest update}
\usage{
data_vintage(id)
}
\arguments{
\item{id}{string, Exact table id}
}
\value{
named character vector. For original data tables the value will be a scalar.
}
\description{
Vintage of table, i.e. datetime of the latest update
}
