% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{datalock}
\alias{datalock}
\title{Lock and Save Data Locally}
\usage{
datalock(
  x,
  name = NULL,
  path = getOption("robonomistClient.datalock.path"),
  overwrite = getOption("robonomistClient.datalock.overwrite", FALSE)
)
}
\arguments{
\item{x}{The R object to be locked and saved.}

\item{name}{Optional; the name to be used for the saved file.
If NULL, a name is generated using the hash of the expression that defines \code{x}.
If provided, it must be a single non-empty character string.}

\item{path}{Optional; the directory path where the data file is to be saved.
Defaults to the value of the 'robonomistClient.datalock.path' option.
If NULL, the file is saved in the current working directory.}

\item{overwrite}{Logical; whether to overwrite the existing data file.
Defaults to FALSE, meaning that if the file exists, it is read and returned
without overwriting.}
}
\value{
Returns the data \code{x} after saving or reading it from the file.
If \code{overwrite} is TRUE, \code{x} is returned after saving.
If \code{overwrite} is FALSE and the file exists, the contents of the file are returned.
If the file does not exist, \code{x} is saved and then returned.
}
\description{
This function locks a given R object \code{x} by saving it to a file with a \code{.rdlock} extension.
If the file already exists (and \code{overwrite} is FALSE), the existing file is read and its contents returned.
}
\note{
This function uses the 'hash' function from the 'rlang' package to generate a file name
if no name is provided.
}
\examples{
x <- runif(1) |> datalock(name = "example", path = "fixed_data")
y <- runif(1) |> datalock(name = "example", path = "fixed_data")
identical(x, y)

# This will save `y` to a file named "example.rdlock" in the "fixed_data" path.
# If the file "example.rdlock" already exists, `x` will to be evalated but read from the file.

}
