% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{datasources}
\alias{datasources}
\title{List available datasources}
\usage{
datasources(lang = NULL)
}
\arguments{
\item{lang, }{Two-letter language code, e.g. "en" or "sv".}
}
\description{
List available datasources
}
