% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogle.R
\name{ogle}
\alias{ogle}
\title{List all distinct values of categorical variables and summarise other variables}
\usage{
ogle(.data, ...)
}
\arguments{
\item{.data}{A tibble}

\item{...}{<'data-masking'> Optional variables to limit ogling. If omitted, all variables will be ogled.}
}
\value{
List of distinct values for character and factor variables
}
\description{
The \code{ogle} function can be helpful in exploring tibble before filtering them.
}
\examples{
\dontrun{
data("sotkanet/4") \%>\% ogle()
data("sotkanet/4") \%>\% ogle(region)
}
}
