% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogle.R
\name{print_filter}
\alias{print_filter}
\title{Print filter criteria for categorical variables in a tibble}
\usage{
print_filter(.data, ...)
}
\arguments{
\item{.data}{A tibble containing the data.}

\item{...}{<'data-masking'> Optional. Variables to include in the filter expression. If omitted, all categorical variables are used.}
}
\value{
Invisibly returns the original tibble. The filter expression is printed to the console and, if possible, copied to the clipboard.
}
\description{
Generates and prints a filter expression for the categorical variables in a tibble.
If run interactively and the \code{clipr} package is available, the filter expression is also copied to the clipboard for easy use in scripts or reports. This helps quickly create filter conditions for further data analysis.
}
\examples{
\dontrun{
data("sotkanet/4") \%>\% print_filter(region_category, gender)
}
}
