% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogle.R
\name{print_filter}
\alias{print_filter}
\title{Print full filter for categorical variables of a tibble}
\usage{
print_filter(.data, ...)
}
\arguments{
\item{.data}{A tibble}

\item{...}{<'data-masking'> Optional variables to limit the printout. If omitted, will print all variables.}
}
\description{
If \code{clipr} package is installed, the filter will be copied to clipboard.
}
\examples{
\dontrun{
data("sotkanet/4") \%>\% print_filter(region_category, gender)
}
}
