% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robonomistClient-package.R
\docType{package}
\name{robonomistClient-package}
\alias{robonomistClient}
\alias{robonomistClient-package}
\title{robonomistClient: Access data from Robonomist Data Server}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Easy and fast access to various datasources in a unified format via Robonomist Data Server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://robonomist.github.io/robonomistClient/}
  \item \url{https://github.com/robonomist/robonomistClient}
}

}
\author{
\strong{Maintainer}: Juha Itkonen \email{juha@robonomist.com}

}
\keyword{internal}
