% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{set_robonomist_server}
\alias{set_robonomist_server}
\title{Set the hostname of Robonomist Data server}
\usage{
set_robonomist_server(
  hostname = getOption("robonomist.server"),
  access_token = getOption("robonomist.access.token")
)
}
\arguments{
\item{hostname}{character, Set the hostname in format "data.hostname.com"}

\item{access_token, }{character, Bearer token}
}
\description{
Set the hostname of Robonomist Data server
}
